requirements.txt
Flask
pyTelegramBotAPI
python-dotenv
waitress
